% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gom_ml.R
\name{gom_ml}
\alias{gom_ml}
\title{Grade of Membership Mixture Model}
\usage{
gom_ml(
  data.object = NULL,
  initial.K = 2,
  final.K = initial.K,
  gamma.algorithm = c("gradient.1992", "woodbury.1974"),
  initial.gamma = c("equal.values", "random", "pure1", "gamma.object"),
  initial.gamma.object = NULL,
  gamma.fit = TRUE,
  lambda.algorithm = c("gradient.1992", "woodbury.1974"),
  initial.lambda = c("random", "pure1", "equal.values", "lambda.object"),
  initial.lambda.object = NULL,
  lambda.fit = TRUE,
  case.id = NA,
  case.weight = NA,
  internal.var = NULL,
  order.K = TRUE,
  omega.fit = FALSE,
  dec.char = ".",
  MC_iter = 1000
)
}
\arguments{
\item{data.object}{A data frame of data or an object that can be coerced to a data frame.}

\item{initial.K}{An integer indicating the initial number of pure type probabilities to be estimated.}

\item{final.K}{An integer indicating the last number of pure type probabilities to be estimated.}

\item{gamma.algorithm}{A character that specifies the algorithm to be used to estimate the gamma values.}

\item{initial.gamma}{A character that specifies how the initial gamma values should be specified. If "gamma.object", then initial.gamma.object is needed.}

\item{initial.gamma.object}{A data frame with the initial gamma values for each observation.}

\item{gamma.fit}{A logical value indicating if the gamma values need to be estimated.}

\item{lambda.algorithm}{A character that specifies the algorithm to be used to estimate the lambda values.}

\item{initial.lambda}{A character that specifies how the initial lambda values should be specified. If "lambda.object", then initial.lambda.object is needed.}

\item{initial.lambda.object}{An array with the initial lambda values for each category of each variable.}

\item{lambda.fit}{A logical value indicating if the lambda values need to be estimated.}

\item{case.id}{A character with the name of the variable with the ID of each observation.}

\item{case.weight}{A character with the name of the variable with the weight of each observation.}

\item{internal.var}{A character vector with the name of the variables to be used by the model.}

\item{order.K}{A logical value indicating if the lambdas should be organized with 99 percent confidence interval.}

\item{omega.fit}{A logical value specifying if the model should generate missing patterns of the data before compiling.}

\item{dec.char}{A character with the decimal symbol to be used in the output files.}

\item{MC_iter}{The number of iterations to be used within the Monte Carlo Estimation.}
}
\value{
gom_ml saves two files, one containing a table with unique data used by the model with the grades of membership, and another
with the initial and final pure type probabilities and some summary statistics. The gom_ml function also returns a list with information about
the results for each number of pure types specified by the user, each of them with following components:

\describe{
  \item{Gik}{A data frame with the grades of membership of each ordered unique pattern of the given data.}
  \item{Pkjl}{An array with dimensions equal to the number of pure types, variables and categories, respectively, with the pure type probabilities.}
  \item{Likelihood}{The maximum log-likelihood achieved by the model.}
  \item{AIC}{The Akaike Information Criterion.}
  \item{Table}{A table with the lambda results organized by variables and their categories.}
}
}
\description{
This function takes a data object and creates a fixed, small set of pure types, each characterized
by the probabilities of answering a category of each question, along with a set of scores (or grades
of membership) that measure each subject’s degree of similarity to each pure type.
}
