% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.gom_bayes.R
\name{print.gom_bayes}
\alias{print.gom_bayes}
\title{Print for \code{gom_bayes} objects}
\usage{
\method{print}{gom_bayes}(x, ...)
}
\arguments{
\item{x}{An object with class \emph{gom_bayes}.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
No return value.
}
\description{
Print method for an objects of class \emph{gom_bayes}.
}
\details{
Prints a descriptive table with the posterior lambdas organized by variables and their categories.
}
\examples{

data <- data.frame(x1 = round(stats::runif(n = 50, 1, 2), 0),
                   x2 = round(stats::runif(n = 50, 1, 3), 0),
                   x3 = round(stats::runif(n = 50, 1, 4), 0))

model <- gom::gom_bayes(data, ntypes = 2, ngibbs = 250, burnin = 250)

print(model)

}
