% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gom_bayes.R
\name{gom_bayes}
\alias{gom_bayes}
\title{Title}
\usage{
gom_bayes(
  dados,
  ntypes = 2,
  alpha = "",
  burnin = 1000,
  ngibbs = 1000,
  omega = 50,
  eta = 10,
  tau = 2,
  beta = 2,
  gomscores = "g"
)
}
\arguments{
\item{dados}{A data frame with the categorical variables to be used be the model.}

\item{ntypes}{An integer indicating the number of pure type probabilities to be estimated.}

\item{alpha}{An array with the pure type probabilities. If specified the model will not try to estimate alpha and will only use the array that was provided by the user.}

\item{burnin}{Number of iterations to be made until the markov chain achieves a stationary distribution}

\item{ngibbs}{Number of iterations to be made after the markov chain achieves a stationary distribution}

\item{omega}{The tuning parameter for the Metropolis–Hastings step.}

\item{eta}{The tuning parameter of the conditional Dirichlet distribution of Xi.}

\item{tau}{The shape parameter of the prior Gamma distribution of alpha.}

\item{beta}{The inverse scale parameter of the prior Gamma distribution of alpha.}

\item{gomscores}{Prefix of the gamma column names.}
}
\value{
A list with the posterior distribution of gamma, lambda, Xi and a latent variable z
}
\description{
Title
}
