% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gom_bayes.R
\name{gom_bayes}
\alias{gom_bayes}
\title{Bayesian Grade of Membership Mixture Model}
\usage{
gom_bayes(
  data,
  ntypes = 2,
  alpha = "",
  burnin = 250,
  ngibbs = 250,
  omega = 50,
  eta = 10,
  tau = 2,
  beta = 2,
  gomscores = "g"
)
}
\arguments{
\item{data}{A data frame with the categorical variables to be used be the model.}

\item{ntypes}{An integer indicating the number of pure type probabilities to be estimated.}

\item{alpha}{An array with the pure type probabilities. If specified, the model will not estimate
alpha and will only use the array provided by user.}

\item{burnin}{Number of iterations for the Markov chain achieve a stationary distribution.}

\item{ngibbs}{Number of iterations after the Markov chain achieved a stationary distribution.}

\item{omega}{The tuning parameter for the Metropolis–Hastings step.}

\item{eta}{The tuning parameter of the conditional Dirichlet distribution of Xi.}

\item{tau}{The shape parameter of the prior Gamma distribution for alpha.}

\item{beta}{The inverse scale parameter of the prior Gamma distribution for alpha.}

\item{gomscores}{Prefix for the gamma column names.}
}
\value{
An object of class \emph{gom_bayes} with the posterior distributions of gamma, lambda, Xi, and alpha.
}
\description{
This function takes a data object and creates a joint posterior distribution of pure type probabilities, which
characterize a small set of extreme profiles, along with the posterior distribution of the grade of membership
estimates and the posterior distribution of the Dirichlet distribution hyperparameters.
}
\examples{
\dontrun{
data <- data.frame(x1 = round(stats::runif(n = 500, 1, 2), 0),
                   x2 = round(stats::runif(n = 500, 1, 3), 0),
                   x3 = round(stats::runif(n = 500, 1, 4), 0),
                   x4 = round(stats::runif(n = 500, 1, 5), 0))

gom_bayes(data, ntypes = 2, ngibbs = 250, burnin = 250)
}
}
