% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.gom_ml.R
\name{print.gom_ml}
\alias{print.gom_ml}
\title{Print for \code{gom_ml} objects}
\usage{
\method{print}{gom_ml}(x, ...)
}
\arguments{
\item{x}{An object with class \emph{gom_ml}.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print objects with class \emph{gom_ml}.
}
\examples{
\dontrun{
data <- data.frame(x1 = round(stats::runif(n = 500, 1, 2), 0),
                   x2 = round(stats::runif(n = 500, 1, 3), 0),
                   x3 = round(stats::runif(n = 500, 1, 4), 0),
                   x4 = round(stats::runif(n = 500, 1, 5), 0),
                   Id = 1:500)

model <- gom_ml(data.object = data, case.id = "Id", initial.lambda = "random")

print(model)
}
}
